Protection Resources with Authorization

+++ Mean.js/MsgBoardApp/backend/routes/posts.js
@@ -38,7 +38,8 @@ router.post(
             title: req.body.title,
             content: req.body.content,
             type: req.body.type,
-            image: url + '/images/' + req.file.filename
+            image: url + '/images/' + req.file.filename,
+            creator: req.userData.userId
         });
         post.save().then(createdPost => {
             res.status(201).json({
@@ -69,9 +70,12 @@ router.put(
             type: req.body.type,
             image: imagePath
         });
-        // console.log(post);
-        Post.updateOne({ _id: req.params.id }, post).then(result => {
-            res.status(200).json({ message: "Update successful!" });
+        Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+            if (result.nModified > 0) {
+                res.status(200).json({ message: "Update Successful!" });
+            } else {
+                res.status(401).json({ message: "Not authorized" });
+            }
         });
     });
 
@@ -108,8 +112,12 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        res.status(200).json({ message: "Post deleted!" });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: "Deletions Successful!" });
+        } else {
+            res.status(401).json({ message: "Not authorized" });
+        }
     });
 });
